Checks for self.markets keys before requesting market ids
The clause "!self.markets.length" is the condition used for market ids lazy loading didn't work because self.markets is an Object (not an array). Now counting Object keys before sending a new request to the server. This is not a robust fix, but it will work if "self.markets" is used only for storing market ids mapping.